fix(casino) exploit of transfer money during game
A user was able to transfer their betting money if they didn't like the
hand they were giving by just giving the money to another account
while the game was running.
This made the payout fail when subtracting money.

Therefore making bailing out of a game an official command but with
casino rules of not having hit a single card.

Also to ensure that nobody loses money give everyone their money
back when reloading and notify the bot owner if something goes wrong
during the back transfer of the bet

closes #36