feat(thread) create thread commands
This cog and its commands are for making "thread-like" channels where
people can join to have longer spoiler discussions, this should help to
reduce the amount of needed spoiler tags within discussing a current
episode.
The channel will delete itself after a certain amount of inactivity and
then gets archived as an text file and attached to the thread join
message.